chore: remediate full-repository code-review findings
Applies fixes from a line-by-line review of the entire repository
(2,455 source/doc files). ~290 Critical/High/Medium findings fixed
across the native Go transform, the typia runtime, @typia/interface,
@typia/utils, the adapter packages, the test suites, website, benchmark
and examples.

Highlights:
- native: `typia generate` now emits the transformed output instead of
  copying the input file; deterministic emit order (sorted map
  iteration) in the Random/Protobuf/MetadataComponents programmers;
  corrected transformer error codes; inverted validation gates fixed in
  the Json/Llm application programmers.
- typia/src: `_isTypeFloat` MINIMUM constant, `_isFormatByte` multiline
  bypass, `_randomArray` unbounded unique-items loop, `_randomFormat*`
  operator-precedence bug, `_isUniqueItems` array length check.
- @typia/utils: OpenAPI up/down-grader bugs (oneOf const loss, filter
  index bug, dropped @default), validator NaN/exclusive-bound fixes,
  NamingConvention empty-string crash.
- tests: export-name/file-name collisions fixed (DynamicExecutor
  discovery), weakened assertions strengthened, CommentTagPattern empty
  @pattern given a real regex.
- website: broken deploy script, doc-sample compile errors, stale links
  and pre-namespace API usages.

Verification: `pnpm build` and `pnpm test` (every tests/test-* workspace
plus `go test ./...`) pass locally.

Deferred to follow-up PRs (genuine findings whose fix surfaces
pre-existing latent defects): the test-error harness `return true`
loop, the test-typia-automated non-create matrix gate, and the
test-utils-automated commented-out spoiler assertion.

Native Go source changed, so the `typia` package version is bumped to
13.0.0-dev.20260521.2 (AGENTS.md section 2.4).

+++ packages/utils/src/validators/internal/OpenApiIntegerValidator.ts
@@ -13,35 +13,35 @@ export namespace OpenApiIntegerValidator {
         ? ctx.value >= ctx.schema.minimum ||
           ctx.report({
             ...ctx,
-            expected: `number & Type<"int32"> & Minimum<${ctx.schema.minimum}>`,
+            expected: `number & Type<"int64"> & Minimum<${ctx.schema.minimum}>`,
           })
         : true,
       ctx.schema.maximum !== undefined
         ? ctx.value <= ctx.schema.maximum ||
           ctx.report({
             ...ctx,
-            expected: `number & Type<"int32"> & Maximum<${ctx.schema.maximum}>`,
+            expected: `number & Type<"int64"> & Maximum<${ctx.schema.maximum}>`,
           })
         : true,
       ctx.schema.exclusiveMinimum !== undefined
         ? ctx.value > ctx.schema.exclusiveMinimum ||
           ctx.report({
             ...ctx,
-            expected: `number & Type<"int32"> & ExclusiveMinimum<${ctx.schema.exclusiveMinimum}>`,
+            expected: `number & Type<"int64"> & ExclusiveMinimum<${ctx.schema.exclusiveMinimum}>`,
           })
         : true,
       ctx.schema.exclusiveMaximum !== undefined
         ? ctx.value < ctx.schema.exclusiveMaximum ||
           ctx.report({
             ...ctx,
-            expected: `number & Type<"int32"> & ExclusiveMaximum<${ctx.schema.exclusiveMaximum}>`,
+            expected: `number & Type<"int64"> & ExclusiveMaximum<${ctx.schema.exclusiveMaximum}>`,
           })
         : true,
       ctx.schema.multipleOf !== undefined
         ? ctx.value % ctx.schema.multipleOf === 0 ||
           ctx.report({
             ...ctx,
-            expected: `number & Type<"int32"> & MultipleOf<${ctx.schema.multipleOf}>`,
+            expected: `number & Type<"int64"> & MultipleOf<${ctx.schema.multipleOf}>`,
           })
         : true,
     ].every((v) => v);
